вынес методы в отдельное апи, добавил плейсхолдер карточке диалога + багфиксы
